Add files via upload
The "Save flashcards" button was modified by adding a background colour and changing its class from `roundArrows` to `saveArrow`. although, background colour for save flashcard button remain thesame as add and view flashcard button to enable consistency of design. In the accompanying CSS, specific styling was added for the `saveArrow` class to differentiate its appearance from the previous and next arrows. The styling includes a different margin, padding, and cursor property. This modification helps visually distinguish the "Save flashcards" button from the navigation arrows in terms of design and functionality.