Add files via upload
The "Share flashcard" functionality was added by introducing a new button with the class `shareButton` and an associated `shareFlashcards` function. This function checks if the Web Share API is supported in the user's browser. If supported, it creates a shareable object with a title, text, and the current URL, then triggers the sharing process. If the Web Share API is not supported, an alert informs the user that sharing is not available in their browser. Finally, an event listener is added to the "Share flashcard" button to execute the `shareFlashcards` function when clicked.