Add files via upload
The share flashcard functionality was modified by adding a custom "Super Share" function using the Clipboard API. If the Web Share API is not supported, the Super Share function is used as a fall back. This enables users to copy flashcard content to the clipboard for sharing. Additionally, a Web Share API function was implemented for a more seamless sharing experience if supported by the browser. The modal overlay and container were also introduced for displaying flashcard information.