Add files via upload
To add audio, I included an `<audio>` element in the HTML's `<head>` section, linking it to a sound file named "flipcard.mp3." This element with the id "flipSound" serves as the audio source. In the JavaScript file "script.js," the function `playFlipSound()` triggers the playback when the "View Answer" button is clicked.